Crate nvme

Structs

ActiveFwInfo

CriticalWarning

EndurGrpCritWarning

ErrLogEntry

FixedStr

FwSlotLog

IdCtrl

IdNmsp

ParamErrLoc

Reserved

SmartLog

StatusField
This includes the phase tag which is inconsistent with the specification. This is done so it is aligned properly.

Enums

CmdSpecificStatus

GenericStatus

MadIntegrityStatus

PathRelatedStatus

StatusCode

StatusCodeType

Traits

FromBytes

TransmuteSafe
Marker trait to indicate that struct can safely be transmuted or cast from bytes without any undefined behavior. This means any contained enums fill their entire repr space. Also, this most likely means the struct is packed.