nvme::ds::controller::properties::aqa

Struct AdminQueueAttributes

pub struct AdminQueueAttributes { /* private fields */ }

Implementations

impl AdminQueueAttributes

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl AdminQueueAttributes

pub const fn into_bytes(self) -> [u8; 4]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 4]) -> Self

Converts the given bytes directly into the bitfield struct.

impl AdminQueueAttributes

pub fn submission_queue_size(&self) -> <B12 as Specifier>::InOut

Returns the value of submission_queue_size.

pub fn submission_queue_size_or_err( &self, ) -> Result<<B12 as Specifier>::InOut, InvalidBitPattern<<B12 as Specifier>::Bytes>>

Returns the value of submission_queue_size.

#Errors

If the returned value contains an invalid bit pattern for submission_queue_size.

pub fn with_submission_queue_size( self, new_val: <B12 as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of submission_queue_size set to the given value.

#Panics

If the given value is out of bounds for submission_queue_size.

pub fn with_submission_queue_size_checked( self, new_val: <B12 as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of submission_queue_size set to the given value.

#Errors

If the given value is out of bounds for submission_queue_size.

pub fn set_submission_queue_size(&mut self, new_val: <B12 as Specifier>::InOut)

Sets the value of submission_queue_size to the given value.

#Panics

If the given value is out of bounds for submission_queue_size.

pub fn set_submission_queue_size_checked( &mut self, new_val: <B12 as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of submission_queue_size to the given value.

#Errors

If the given value is out of bounds for submission_queue_size.

pub fn completion_queue_size(&self) -> <B12 as Specifier>::InOut

Returns the value of completion_queue_size.

pub fn completion_queue_size_or_err( &self, ) -> Result<<B12 as Specifier>::InOut, InvalidBitPattern<<B12 as Specifier>::Bytes>>

Returns the value of completion_queue_size.

#Errors

If the returned value contains an invalid bit pattern for completion_queue_size.

pub fn with_completion_queue_size( self, new_val: <B12 as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of completion_queue_size set to the given value.

#Panics

If the given value is out of bounds for completion_queue_size.

pub fn with_completion_queue_size_checked( self, new_val: <B12 as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of completion_queue_size set to the given value.

#Errors

If the given value is out of bounds for completion_queue_size.

pub fn set_completion_queue_size(&mut self, new_val: <B12 as Specifier>::InOut)

Sets the value of completion_queue_size to the given value.

#Panics

If the given value is out of bounds for completion_queue_size.

pub fn set_completion_queue_size_checked( &mut self, new_val: <B12 as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of completion_queue_size to the given value.

#Errors

If the given value is out of bounds for completion_queue_size.