nvme::ds::controller::properties::capabilities

Struct ControllerCap

pub struct ControllerCap { /* private fields */ }

Implementations

impl ControllerCap

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl ControllerCap

pub const fn into_bytes(self) -> [u8; 8]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 8]) -> Self

Converts the given bytes directly into the bitfield struct.

impl ControllerCap

pub fn max_queue_entries(&self) -> <u16 as Specifier>::InOut

Returns the value of max_queue_entries.

pub fn max_queue_entries_or_err( &self, ) -> Result<<u16 as Specifier>::InOut, InvalidBitPattern<<u16 as Specifier>::Bytes>>

Returns the value of max_queue_entries.

#Errors

If the returned value contains an invalid bit pattern for max_queue_entries.

pub fn contiguous_queues_required(&self) -> <bool as Specifier>::InOut

Returns the value of contiguous_queues_required.

pub fn contiguous_queues_required_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of contiguous_queues_required.

#Errors

If the returned value contains an invalid bit pattern for contiguous_queues_required.

pub fn timeout(&self) -> <HalfSeconds as Specifier>::InOut

Returns the value of timeout.

pub fn timeout_or_err( &self, ) -> Result<<HalfSeconds as Specifier>::InOut, InvalidBitPattern<<HalfSeconds as Specifier>::Bytes>>

Returns the value of timeout.

#Errors

If the returned value contains an invalid bit pattern for timeout.

pub fn doorbell_stride(&self) -> <B4 as Specifier>::InOut

Returns the value of doorbell_stride.

pub fn doorbell_stride_or_err( &self, ) -> Result<<B4 as Specifier>::InOut, InvalidBitPattern<<B4 as Specifier>::Bytes>>

Returns the value of doorbell_stride.

#Errors

If the returned value contains an invalid bit pattern for doorbell_stride.

pub fn nvm_subsystem_reset_supported(&self) -> <bool as Specifier>::InOut

Returns the value of nvm_subsystem_reset_supported.

pub fn nvm_subsystem_reset_supported_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of nvm_subsystem_reset_supported.

#Errors

If the returned value contains an invalid bit pattern for nvm_subsystem_reset_supported.

pub fn boot_partition_supported(&self) -> <bool as Specifier>::InOut

Returns the value of boot_partition_supported.

pub fn boot_partition_supported_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of boot_partition_supported.

#Errors

If the returned value contains an invalid bit pattern for boot_partition_supported.

pub fn controller_power_scope( &self, ) -> <ControllerPowerScope as Specifier>::InOut

Returns the value of controller_power_scope.

pub fn controller_power_scope_or_err( &self, ) -> Result<<ControllerPowerScope as Specifier>::InOut, InvalidBitPattern<<ControllerPowerScope as Specifier>::Bytes>>

Returns the value of controller_power_scope.

#Errors

If the returned value contains an invalid bit pattern for controller_power_scope.

pub fn memory_page_sz_min(&self) -> <B4 as Specifier>::InOut

Returns the value of memory_page_sz_min.

pub fn memory_page_sz_min_or_err( &self, ) -> Result<<B4 as Specifier>::InOut, InvalidBitPattern<<B4 as Specifier>::Bytes>>

Returns the value of memory_page_sz_min.

#Errors

If the returned value contains an invalid bit pattern for memory_page_sz_min.

pub fn memory_page_sz_max(&self) -> <B4 as Specifier>::InOut

Returns the value of memory_page_sz_max.

pub fn memory_page_sz_max_or_err( &self, ) -> Result<<B4 as Specifier>::InOut, InvalidBitPattern<<B4 as Specifier>::Bytes>>

Returns the value of memory_page_sz_max.

#Errors

If the returned value contains an invalid bit pattern for memory_page_sz_max.

pub fn persistent_mem_region_supported(&self) -> <bool as Specifier>::InOut

Returns the value of persistent_mem_region_supported.

pub fn persistent_mem_region_supported_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of persistent_mem_region_supported.

#Errors

If the returned value contains an invalid bit pattern for persistent_mem_region_supported.

pub fn controller_mem_buffer_supported(&self) -> <bool as Specifier>::InOut

Returns the value of controller_mem_buffer_supported.

pub fn controller_mem_buffer_supported_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of controller_mem_buffer_supported.

#Errors

If the returned value contains an invalid bit pattern for controller_mem_buffer_supported.

pub fn nvm_subsystem_shutdown_supported(&self) -> <bool as Specifier>::InOut

Returns the value of nvm_subsystem_shutdown_supported.

pub fn nvm_subsystem_shutdown_supported_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of nvm_subsystem_shutdown_supported.

#Errors

If the returned value contains an invalid bit pattern for nvm_subsystem_shutdown_supported.

pub fn controller_ready_modes_supported(&self) -> <B2 as Specifier>::InOut

Returns the value of controller_ready_modes_supported.

pub fn controller_ready_modes_supported_or_err( &self, ) -> Result<<B2 as Specifier>::InOut, InvalidBitPattern<<B2 as Specifier>::Bytes>>

Returns the value of controller_ready_modes_supported.

#Errors

If the returned value contains an invalid bit pattern for controller_ready_modes_supported.

impl ControllerCap

pub fn controller_ready_with_media_support(&self) -> bool

pub fn controller_ready_independent_of_media_support(&self) -> bool

pub fn arbitration_urgent_with_wrr(&self) -> bool

pub fn arbitration_vendor_specific(&self) -> bool

pub fn supports_nvm_command_set(&self) -> bool

pub fn supports_more_io_command_sets(&self) -> bool

pub fn doorbell_stride_bytes(&self) -> usize

pub fn memory_page_sz_min_bytes(&self) -> usize

pub fn memory_page_sz_max_bytes(&self) -> usize