Struct nvme::ds::identify::controller::ControllerAttributes

pub struct ControllerAttributes { /* private fields */ }

Implementations

impl ControllerAttributes

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl ControllerAttributes

pub const fn into_bytes(self) -> [u8; 4]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 4]) -> Self

Converts the given bytes directly into the bitfield struct.

impl ControllerAttributes

pub fn host_id(&self) -> <B1 as Specifier>::InOut

Returns the value of host_id.

pub fn host_id_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of host_id.

#Errors

If the returned value contains an invalid bit pattern for host_id.

pub fn non_op_power_state_permissive_mode(&self) -> <B1 as Specifier>::InOut

Returns the value of non_op_power_state_permissive_mode.

pub fn non_op_power_state_permissive_mode_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of non_op_power_state_permissive_mode.

#Errors

If the returned value contains an invalid bit pattern for non_op_power_state_permissive_mode.

pub fn nvm_sets(&self) -> <B1 as Specifier>::InOut

Returns the value of nvm_sets.

pub fn nvm_sets_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of nvm_sets.

#Errors

If the returned value contains an invalid bit pattern for nvm_sets.

pub fn read_recovery_levels(&self) -> <B1 as Specifier>::InOut

Returns the value of read_recovery_levels.

pub fn read_recovery_levels_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of read_recovery_levels.

#Errors

If the returned value contains an invalid bit pattern for read_recovery_levels.

pub fn endurance_groups(&self) -> <B1 as Specifier>::InOut

Returns the value of endurance_groups.

pub fn endurance_groups_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of endurance_groups.

#Errors

If the returned value contains an invalid bit pattern for endurance_groups.

pub fn predictable_latency_mode(&self) -> <B1 as Specifier>::InOut

Returns the value of predictable_latency_mode.

pub fn predictable_latency_mode_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of predictable_latency_mode.

#Errors

If the returned value contains an invalid bit pattern for predictable_latency_mode.

pub fn traffic_based_keep_alive(&self) -> <B1 as Specifier>::InOut

Returns the value of traffic_based_keep_alive.

pub fn traffic_based_keep_alive_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of traffic_based_keep_alive.

#Errors

If the returned value contains an invalid bit pattern for traffic_based_keep_alive.

pub fn namespace_granularity(&self) -> <B1 as Specifier>::InOut

Returns the value of namespace_granularity.

pub fn namespace_granularity_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of namespace_granularity.

#Errors

If the returned value contains an invalid bit pattern for namespace_granularity.

pub fn sq_associations(&self) -> <B1 as Specifier>::InOut

Returns the value of sq_associations.

pub fn sq_associations_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of sq_associations.

#Errors

If the returned value contains an invalid bit pattern for sq_associations.

pub fn uuid_list(&self) -> <B1 as Specifier>::InOut

Returns the value of uuid_list.

pub fn uuid_list_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of uuid_list.

#Errors

If the returned value contains an invalid bit pattern for uuid_list.

pub fn multi_domain_subsystem(&self) -> <B1 as Specifier>::InOut

Returns the value of multi_domain_subsystem.

pub fn multi_domain_subsystem_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of multi_domain_subsystem.

#Errors

If the returned value contains an invalid bit pattern for multi_domain_subsystem.

pub fn fixed_domain_capacity_management(&self) -> <B1 as Specifier>::InOut

Returns the value of fixed_domain_capacity_management.

pub fn fixed_domain_capacity_management_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of fixed_domain_capacity_management.

#Errors

If the returned value contains an invalid bit pattern for fixed_domain_capacity_management.

pub fn variable_capacity_management(&self) -> <B1 as Specifier>::InOut

Returns the value of variable_capacity_management.

pub fn variable_capacity_management_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of variable_capacity_management.

#Errors

If the returned value contains an invalid bit pattern for variable_capacity_management.

pub fn delete_endurance_group(&self) -> <B1 as Specifier>::InOut

Returns the value of delete_endurance_group.

pub fn delete_endurance_group_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of delete_endurance_group.

#Errors

If the returned value contains an invalid bit pattern for delete_endurance_group.

pub fn delete_nvm_set(&self) -> <B1 as Specifier>::InOut

Returns the value of delete_nvm_set.

pub fn delete_nvm_set_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of delete_nvm_set.

#Errors

If the returned value contains an invalid bit pattern for delete_nvm_set.

pub fn extended_lba_formats_supported(&self) -> <B1 as Specifier>::InOut

Returns the value of extended_lba_formats_supported.

pub fn extended_lba_formats_supported_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of extended_lba_formats_supported.

#Errors

If the returned value contains an invalid bit pattern for extended_lba_formats_supported.