Struct nvme::ds::identify::controller::OptionalAdminCommandSupport

pub struct OptionalAdminCommandSupport { /* private fields */ }

Implementations

impl OptionalAdminCommandSupport

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl OptionalAdminCommandSupport

pub const fn into_bytes(self) -> [u8; 2]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 2]) -> Self

Converts the given bytes directly into the bitfield struct.

impl OptionalAdminCommandSupport

pub fn security_send_and_recv(&self) -> <B1 as Specifier>::InOut

Returns the value of security_send_and_recv.

pub fn security_send_and_recv_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of security_send_and_recv.

#Errors

If the returned value contains an invalid bit pattern for security_send_and_recv.

pub fn format_nvm(&self) -> <B1 as Specifier>::InOut

Returns the value of format_nvm.

pub fn format_nvm_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of format_nvm.

#Errors

If the returned value contains an invalid bit pattern for format_nvm.

pub fn firmware_download_and_commit(&self) -> <B1 as Specifier>::InOut

Returns the value of firmware_download_and_commit.

pub fn firmware_download_and_commit_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of firmware_download_and_commit.

#Errors

If the returned value contains an invalid bit pattern for firmware_download_and_commit.

pub fn namespace_management(&self) -> <B1 as Specifier>::InOut

Returns the value of namespace_management.

pub fn namespace_management_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of namespace_management.

#Errors

If the returned value contains an invalid bit pattern for namespace_management.

pub fn device_self_test(&self) -> <B1 as Specifier>::InOut

Returns the value of device_self_test.

pub fn device_self_test_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of device_self_test.

#Errors

If the returned value contains an invalid bit pattern for device_self_test.

pub fn directives(&self) -> <B1 as Specifier>::InOut

Returns the value of directives.

pub fn directives_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of directives.

#Errors

If the returned value contains an invalid bit pattern for directives.

pub fn nvme_mi_send_and_recv(&self) -> <B1 as Specifier>::InOut

Returns the value of nvme_mi_send_and_recv.

pub fn nvme_mi_send_and_recv_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of nvme_mi_send_and_recv.

#Errors

If the returned value contains an invalid bit pattern for nvme_mi_send_and_recv.

pub fn virtualization_management(&self) -> <B1 as Specifier>::InOut

Returns the value of virtualization_management.

pub fn virtualization_management_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of virtualization_management.

#Errors

If the returned value contains an invalid bit pattern for virtualization_management.

pub fn doorbell_buffer_config(&self) -> <B1 as Specifier>::InOut

Returns the value of doorbell_buffer_config.

pub fn doorbell_buffer_config_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of doorbell_buffer_config.

#Errors

If the returned value contains an invalid bit pattern for doorbell_buffer_config.

pub fn get_lba_status(&self) -> <B1 as Specifier>::InOut

Returns the value of get_lba_status.

pub fn get_lba_status_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of get_lba_status.

#Errors

If the returned value contains an invalid bit pattern for get_lba_status.

pub fn command_and_feature_lockdown(&self) -> <B1 as Specifier>::InOut

Returns the value of command_and_feature_lockdown.

pub fn command_and_feature_lockdown_or_err( &self ) -> Result<<B1 as Specifier>::InOut, InvalidBitPattern<<B1 as Specifier>::Bytes>>

Returns the value of command_and_feature_lockdown.

#Errors

If the returned value contains an invalid bit pattern for command_and_feature_lockdown.