nvme::nvm

Struct ContextAttributes

pub struct ContextAttributes { /* private fields */ }

Implementations

impl ContextAttributes

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl ContextAttributes

pub const fn into_bytes(self) -> [u8; 4]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 4]) -> Self

Converts the given bytes directly into the bitfield struct.

impl ContextAttributes

pub fn access_frequency(&self) -> <AccessFrequency as Specifier>::InOut

Returns the value of access_frequency.

pub fn access_frequency_or_err( &self, ) -> Result<<AccessFrequency as Specifier>::InOut, InvalidBitPattern<<AccessFrequency as Specifier>::Bytes>>

Returns the value of access_frequency.

Errors

If the returned value contains an invalid bit pattern for access_frequency.

pub fn with_access_frequency( self, new_val: <AccessFrequency as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of access_frequency set to the given value.

Panics

If the given value is out of bounds for access_frequency.

pub fn with_access_frequency_checked( self, new_val: <AccessFrequency as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of access_frequency set to the given value.

Errors

If the given value is out of bounds for access_frequency.

pub fn set_access_frequency( &mut self, new_val: <AccessFrequency as Specifier>::InOut, )

Sets the value of access_frequency to the given value.

Panics

If the given value is out of bounds for access_frequency.

pub fn set_access_frequency_checked( &mut self, new_val: <AccessFrequency as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of access_frequency to the given value.

Errors

If the given value is out of bounds for access_frequency.

pub fn access_latency(&self) -> <AccessLatency as Specifier>::InOut

Returns the value of access_latency.

pub fn access_latency_or_err( &self, ) -> Result<<AccessLatency as Specifier>::InOut, InvalidBitPattern<<AccessLatency as Specifier>::Bytes>>

Returns the value of access_latency.

Errors

If the returned value contains an invalid bit pattern for access_latency.

pub fn with_access_latency( self, new_val: <AccessLatency as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of access_latency set to the given value.

Panics

If the given value is out of bounds for access_latency.

pub fn with_access_latency_checked( self, new_val: <AccessLatency as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of access_latency set to the given value.

Errors

If the given value is out of bounds for access_latency.

pub fn set_access_latency( &mut self, new_val: <AccessLatency as Specifier>::InOut, )

Sets the value of access_latency to the given value.

Panics

If the given value is out of bounds for access_latency.

pub fn set_access_latency_checked( &mut self, new_val: <AccessLatency as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of access_latency to the given value.

Errors

If the given value is out of bounds for access_latency.

pub fn seq_read_range(&self) -> <bool as Specifier>::InOut

Returns the value of seq_read_range.

pub fn seq_read_range_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of seq_read_range.

Errors

If the returned value contains an invalid bit pattern for seq_read_range.

pub fn with_seq_read_range(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of seq_read_range set to the given value.

Panics

If the given value is out of bounds for seq_read_range.

pub fn with_seq_read_range_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of seq_read_range set to the given value.

Errors

If the given value is out of bounds for seq_read_range.

pub fn set_seq_read_range(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of seq_read_range to the given value.

Panics

If the given value is out of bounds for seq_read_range.

pub fn set_seq_read_range_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of seq_read_range to the given value.

Errors

If the given value is out of bounds for seq_read_range.

pub fn seq_write_range(&self) -> <bool as Specifier>::InOut

Returns the value of seq_write_range.

pub fn seq_write_range_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of seq_write_range.

Errors

If the returned value contains an invalid bit pattern for seq_write_range.

pub fn with_seq_write_range(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of seq_write_range set to the given value.

Panics

If the given value is out of bounds for seq_write_range.

pub fn with_seq_write_range_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of seq_write_range set to the given value.

Errors

If the given value is out of bounds for seq_write_range.

pub fn set_seq_write_range(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of seq_write_range to the given value.

Panics

If the given value is out of bounds for seq_write_range.

pub fn set_seq_write_range_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of seq_write_range to the given value.

Errors

If the given value is out of bounds for seq_write_range.

pub fn write_prepare(&self) -> <bool as Specifier>::InOut

Returns the value of write_prepare.

pub fn write_prepare_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of write_prepare.

Errors

If the returned value contains an invalid bit pattern for write_prepare.

pub fn with_write_prepare(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of write_prepare set to the given value.

Panics

If the given value is out of bounds for write_prepare.

pub fn with_write_prepare_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of write_prepare set to the given value.

Errors

If the given value is out of bounds for write_prepare.

pub fn set_write_prepare(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of write_prepare to the given value.

Panics

If the given value is out of bounds for write_prepare.

pub fn set_write_prepare_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of write_prepare to the given value.

Errors

If the given value is out of bounds for write_prepare.

pub fn command_access_size(&self) -> <u8 as Specifier>::InOut

Returns the value of command_access_size.

pub fn command_access_size_or_err( &self, ) -> Result<<u8 as Specifier>::InOut, InvalidBitPattern<<u8 as Specifier>::Bytes>>

Returns the value of command_access_size.

Errors

If the returned value contains an invalid bit pattern for command_access_size.

pub fn with_command_access_size(self, new_val: <u8 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of command_access_size set to the given value.

Panics

If the given value is out of bounds for command_access_size.

pub fn with_command_access_size_checked( self, new_val: <u8 as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of command_access_size set to the given value.

Errors

If the given value is out of bounds for command_access_size.

pub fn set_command_access_size(&mut self, new_val: <u8 as Specifier>::InOut)

Sets the value of command_access_size to the given value.

Panics

If the given value is out of bounds for command_access_size.

pub fn set_command_access_size_checked( &mut self, new_val: <u8 as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of command_access_size to the given value.

Errors

If the given value is out of bounds for command_access_size.