Struct nvme::StatusField

pub struct StatusField { /* fields omitted */ }

This includes the phase tag which is inconsistent with the specification. This is done so it is aligned properly.

Implementations

impl StatusField

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl StatusField

pub const fn into_bytes(self) -> [u8; 2]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 2]) -> Self

Converts the given bytes directly into the bitfield struct.

impl StatusField

pub fn phase_tag(&self) -> <bool as Specifier>::InOut

Returns the value of phase_tag.

pub fn phase_tag_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of phase_tag.

#Errors

If the returned value contains an invalid bit pattern for phase_tag.

pub fn with_phase_tag(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of phase_tag set to the given value.

#Panics

If the given value is out of bounds for phase_tag.

pub fn with_phase_tag_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of phase_tag set to the given value.

#Errors

If the given value is out of bounds for phase_tag.

pub fn set_phase_tag(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of phase_tag to the given value.

#Panics

If the given value is out of bounds for phase_tag.

pub fn set_phase_tag_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of phase_tag to the given value.

#Errors

If the given value is out of bounds for phase_tag.

pub fn sc(&self) -> <u8 as Specifier>::InOut

Returns the value of sc.

pub fn sc_or_err(
&self
) -> Result<<u8 as Specifier>::InOut, InvalidBitPattern<<u8 as Specifier>::Bytes>>

Returns the value of sc.

#Errors

If the returned value contains an invalid bit pattern for sc.

pub fn with_sc(self, new_val: <u8 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of sc set to the given value.

#Panics

If the given value is out of bounds for sc.

pub fn with_sc_checked(
self,
new_val: <u8 as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of sc set to the given value.

#Errors

If the given value is out of bounds for sc.

pub fn set_sc(&mut self, new_val: <u8 as Specifier>::InOut)

Sets the value of sc to the given value.

#Panics

If the given value is out of bounds for sc.

pub fn set_sc_checked(
&mut self,
new_val: <u8 as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of sc to the given value.

#Errors

If the given value is out of bounds for sc.

pub fn sct(&self) -> <StatusCodeType as Specifier>::InOut

Returns the value of sct.

pub fn sct_or_err(
&self
) -> Result<<StatusCodeType as Specifier>::InOut, InvalidBitPattern<<StatusCodeType as Specifier>::Bytes>>

Returns the value of sct.

#Errors

If the returned value contains an invalid bit pattern for sct.

pub fn with_sct(self, new_val: <StatusCodeType as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of sct set to the given value.

#Panics

If the given value is out of bounds for sct.

pub fn with_sct_checked(
self,
new_val: <StatusCodeType as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of sct set to the given value.

#Errors

If the given value is out of bounds for sct.

pub fn set_sct(&mut self, new_val: <StatusCodeType as Specifier>::InOut)

Sets the value of sct to the given value.

#Panics

If the given value is out of bounds for sct.

pub fn set_sct_checked(
&mut self,
new_val: <StatusCodeType as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of sct to the given value.

#Errors

If the given value is out of bounds for sct.

pub fn crd(&self) -> <B2 as Specifier>::InOut

Returns the value of crd.

pub fn crd_or_err(
&self
) -> Result<<B2 as Specifier>::InOut, InvalidBitPattern<<B2 as Specifier>::Bytes>>

Returns the value of crd.

#Errors

If the returned value contains an invalid bit pattern for crd.

pub fn with_crd(self, new_val: <B2 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of crd set to the given value.

#Panics

If the given value is out of bounds for crd.

pub fn with_crd_checked(
self,
new_val: <B2 as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of crd set to the given value.

#Errors

If the given value is out of bounds for crd.

pub fn set_crd(&mut self, new_val: <B2 as Specifier>::InOut)

Sets the value of crd to the given value.

#Panics

If the given value is out of bounds for crd.

pub fn set_crd_checked(
&mut self,
new_val: <B2 as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of crd to the given value.

#Errors

If the given value is out of bounds for crd.

pub fn more(&self) -> <bool as Specifier>::InOut

Returns the value of more.

pub fn more_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of more.

#Errors

If the returned value contains an invalid bit pattern for more.

pub fn with_more(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of more set to the given value.

#Panics

If the given value is out of bounds for more.

pub fn with_more_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of more set to the given value.

#Errors

If the given value is out of bounds for more.

pub fn set_more(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of more to the given value.

#Panics

If the given value is out of bounds for more.

pub fn set_more_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of more to the given value.

#Errors

If the given value is out of bounds for more.

pub fn dnr(&self) -> <bool as Specifier>::InOut

Returns the value of dnr.

pub fn dnr_or_err(
&self
) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of dnr.

#Errors

If the returned value contains an invalid bit pattern for dnr.

pub fn with_dnr(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of dnr set to the given value.

#Panics

If the given value is out of bounds for dnr.

pub fn with_dnr_checked(
self,
new_val: <bool as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of dnr set to the given value.

#Errors

If the given value is out of bounds for dnr.

pub fn set_dnr(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of dnr to the given value.

#Panics

If the given value is out of bounds for dnr.

pub fn set_dnr_checked(
&mut self,
new_val: <bool as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of dnr to the given value.

#Errors

If the given value is out of bounds for dnr.

impl StatusField

pub fn status_code(&self) -> StatusCode

pub fn successful(&self) -> bool