nvme::ds::controller::properties::config
Struct ControllerConfig

pub struct ControllerConfig { /* private fields */ }

Implementations

impl ControllerConfig

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl ControllerConfig

pub const fn into_bytes(self) -> [u8; 4]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 4]) -> Self

Converts the given bytes directly into the bitfield struct.

impl ControllerConfig

pub fn enable(&self) -> <bool as Specifier>::InOut

Returns the value of enable.

pub fn enable_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of enable.

#Errors

If the returned value contains an invalid bit pattern for enable.

pub fn with_enable(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of enable set to the given value.

#Panics

If the given value is out of bounds for enable.

pub fn with_enable_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of enable set to the given value.

#Errors

If the given value is out of bounds for enable.

pub fn set_enable(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of enable to the given value.

#Panics

If the given value is out of bounds for enable.

pub fn set_enable_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of enable to the given value.

#Errors

If the given value is out of bounds for enable.

pub fn io_command_set_selected(&self) -> <IOCommandSet as Specifier>::InOut

Returns the value of io_command_set_selected.

pub fn io_command_set_selected_or_err( &self, ) -> Result<<IOCommandSet as Specifier>::InOut, InvalidBitPattern<<IOCommandSet as Specifier>::Bytes>>

Returns the value of io_command_set_selected.

#Errors

If the returned value contains an invalid bit pattern for io_command_set_selected.

pub fn with_io_command_set_selected( self, new_val: <IOCommandSet as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of io_command_set_selected set to the given value.

#Panics

If the given value is out of bounds for io_command_set_selected.

pub fn with_io_command_set_selected_checked( self, new_val: <IOCommandSet as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of io_command_set_selected set to the given value.

#Errors

If the given value is out of bounds for io_command_set_selected.

pub fn set_io_command_set_selected( &mut self, new_val: <IOCommandSet as Specifier>::InOut, )

Sets the value of io_command_set_selected to the given value.

#Panics

If the given value is out of bounds for io_command_set_selected.

pub fn set_io_command_set_selected_checked( &mut self, new_val: <IOCommandSet as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of io_command_set_selected to the given value.

#Errors

If the given value is out of bounds for io_command_set_selected.

pub fn mem_page_size(&self) -> <B4 as Specifier>::InOut

Returns the value of mem_page_size.

pub fn mem_page_size_or_err( &self, ) -> Result<<B4 as Specifier>::InOut, InvalidBitPattern<<B4 as Specifier>::Bytes>>

Returns the value of mem_page_size.

#Errors

If the returned value contains an invalid bit pattern for mem_page_size.

pub fn with_mem_page_size(self, new_val: <B4 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of mem_page_size set to the given value.

#Panics

If the given value is out of bounds for mem_page_size.

pub fn with_mem_page_size_checked( self, new_val: <B4 as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of mem_page_size set to the given value.

#Errors

If the given value is out of bounds for mem_page_size.

pub fn set_mem_page_size(&mut self, new_val: <B4 as Specifier>::InOut)

Sets the value of mem_page_size to the given value.

#Panics

If the given value is out of bounds for mem_page_size.

pub fn set_mem_page_size_checked( &mut self, new_val: <B4 as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of mem_page_size to the given value.

#Errors

If the given value is out of bounds for mem_page_size.

pub fn arbitration_mechanism( &self, ) -> <ArbitrationMechanism as Specifier>::InOut

Returns the value of arbitration_mechanism.

pub fn arbitration_mechanism_or_err( &self, ) -> Result<<ArbitrationMechanism as Specifier>::InOut, InvalidBitPattern<<ArbitrationMechanism as Specifier>::Bytes>>

Returns the value of arbitration_mechanism.

#Errors

If the returned value contains an invalid bit pattern for arbitration_mechanism.

pub fn with_arbitration_mechanism( self, new_val: <ArbitrationMechanism as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of arbitration_mechanism set to the given value.

#Panics

If the given value is out of bounds for arbitration_mechanism.

pub fn with_arbitration_mechanism_checked( self, new_val: <ArbitrationMechanism as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of arbitration_mechanism set to the given value.

#Errors

If the given value is out of bounds for arbitration_mechanism.

pub fn set_arbitration_mechanism( &mut self, new_val: <ArbitrationMechanism as Specifier>::InOut, )

Sets the value of arbitration_mechanism to the given value.

#Panics

If the given value is out of bounds for arbitration_mechanism.

pub fn set_arbitration_mechanism_checked( &mut self, new_val: <ArbitrationMechanism as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of arbitration_mechanism to the given value.

#Errors

If the given value is out of bounds for arbitration_mechanism.

pub fn shutdown_notification( &self, ) -> <ShutdownNotification as Specifier>::InOut

Returns the value of shutdown_notification.

pub fn shutdown_notification_or_err( &self, ) -> Result<<ShutdownNotification as Specifier>::InOut, InvalidBitPattern<<ShutdownNotification as Specifier>::Bytes>>

Returns the value of shutdown_notification.

#Errors

If the returned value contains an invalid bit pattern for shutdown_notification.

pub fn with_shutdown_notification( self, new_val: <ShutdownNotification as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of shutdown_notification set to the given value.

#Panics

If the given value is out of bounds for shutdown_notification.

pub fn with_shutdown_notification_checked( self, new_val: <ShutdownNotification as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of shutdown_notification set to the given value.

#Errors

If the given value is out of bounds for shutdown_notification.

pub fn set_shutdown_notification( &mut self, new_val: <ShutdownNotification as Specifier>::InOut, )

Sets the value of shutdown_notification to the given value.

#Panics

If the given value is out of bounds for shutdown_notification.

pub fn set_shutdown_notification_checked( &mut self, new_val: <ShutdownNotification as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of shutdown_notification to the given value.

#Errors

If the given value is out of bounds for shutdown_notification.

pub fn io_submission_queue_entry_size(&self) -> <B4 as Specifier>::InOut

Returns the value of io_submission_queue_entry_size.

pub fn io_submission_queue_entry_size_or_err( &self, ) -> Result<<B4 as Specifier>::InOut, InvalidBitPattern<<B4 as Specifier>::Bytes>>

Returns the value of io_submission_queue_entry_size.

#Errors

If the returned value contains an invalid bit pattern for io_submission_queue_entry_size.

pub fn with_io_submission_queue_entry_size( self, new_val: <B4 as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of io_submission_queue_entry_size set to the given value.

#Panics

If the given value is out of bounds for io_submission_queue_entry_size.

pub fn with_io_submission_queue_entry_size_checked( self, new_val: <B4 as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of io_submission_queue_entry_size set to the given value.

#Errors

If the given value is out of bounds for io_submission_queue_entry_size.

pub fn set_io_submission_queue_entry_size( &mut self, new_val: <B4 as Specifier>::InOut, )

Sets the value of io_submission_queue_entry_size to the given value.

#Panics

If the given value is out of bounds for io_submission_queue_entry_size.

pub fn set_io_submission_queue_entry_size_checked( &mut self, new_val: <B4 as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of io_submission_queue_entry_size to the given value.

#Errors

If the given value is out of bounds for io_submission_queue_entry_size.

pub fn io_completion_queue_entry_size(&self) -> <B4 as Specifier>::InOut

Returns the value of io_completion_queue_entry_size.

pub fn io_completion_queue_entry_size_or_err( &self, ) -> Result<<B4 as Specifier>::InOut, InvalidBitPattern<<B4 as Specifier>::Bytes>>

Returns the value of io_completion_queue_entry_size.

#Errors

If the returned value contains an invalid bit pattern for io_completion_queue_entry_size.

pub fn with_io_completion_queue_entry_size( self, new_val: <B4 as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of io_completion_queue_entry_size set to the given value.

#Panics

If the given value is out of bounds for io_completion_queue_entry_size.

pub fn with_io_completion_queue_entry_size_checked( self, new_val: <B4 as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of io_completion_queue_entry_size set to the given value.

#Errors

If the given value is out of bounds for io_completion_queue_entry_size.

pub fn set_io_completion_queue_entry_size( &mut self, new_val: <B4 as Specifier>::InOut, )

Sets the value of io_completion_queue_entry_size to the given value.

#Panics

If the given value is out of bounds for io_completion_queue_entry_size.

pub fn set_io_completion_queue_entry_size_checked( &mut self, new_val: <B4 as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of io_completion_queue_entry_size to the given value.

#Errors

If the given value is out of bounds for io_completion_queue_entry_size.

pub fn controller_ready_independent_of_media_enable( &self, ) -> <bool as Specifier>::InOut

Returns the value of controller_ready_independent_of_media_enable.

pub fn controller_ready_independent_of_media_enable_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of controller_ready_independent_of_media_enable.

#Errors

If the returned value contains an invalid bit pattern for controller_ready_independent_of_media_enable.

pub fn with_controller_ready_independent_of_media_enable( self, new_val: <bool as Specifier>::InOut, ) -> Self

Returns a copy of the bitfield with the value of controller_ready_independent_of_media_enable set to the given value.

#Panics

If the given value is out of bounds for controller_ready_independent_of_media_enable.

pub fn with_controller_ready_independent_of_media_enable_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of controller_ready_independent_of_media_enable set to the given value.

#Errors

If the given value is out of bounds for controller_ready_independent_of_media_enable.

pub fn set_controller_ready_independent_of_media_enable( &mut self, new_val: <bool as Specifier>::InOut, )

Sets the value of controller_ready_independent_of_media_enable to the given value.

#Panics

If the given value is out of bounds for controller_ready_independent_of_media_enable.

pub fn set_controller_ready_independent_of_media_enable_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of controller_ready_independent_of_media_enable to the given value.

#Errors

If the given value is out of bounds for controller_ready_independent_of_media_enable.